+++ changelog.md
@@ -46,3 +46,4 @@
 - Systematic catching of errors in request handlers.
 - Subtracted content logic for custom file into its own function getCustomFileContent.
 - Modified serveDownloadFile to expect parts to be a comma seperated string. 
+- Updated serveDownloadFile to get source files from GitHub.